Tests

+++ src/main/java/jv/records/Data.java
@@ -1,2 +1,2 @@
 package jv.records;
-public record Data(String term, int doc_id, double value){}
+public record Data(String term, long doc_id, double value){}
